+++ src/ServiceControl.Monitoring/HostApplicationBuilderExtensions.cs
@@ -38,7 +38,6 @@ public static class HostApplicationBuilderExtensions
 
         var services = hostBuilder.Services;
         services.AddSingleton(settings);
-        services.AddSingleton<LicenseCheckFeatureStartup>();
         services.AddSingleton<EndpointRegistry>();
         services.AddSingleton<MessageTypeRegistry>();
         services.AddSingleton<EndpointInstanceActivityTracker>();
@@ -61,11 +60,13 @@ public static class HostApplicationBuilderExtensions
         // directly and to make things more complex of course the order of registration still matters ;)
         services.AddSingleton(provider => new Lazy<IMessageDispatcher>(provider.GetRequiredService<IMessageDispatcher>));
 
+        services.AddLicenseCheck();
+
         ConfigureEndpoint(endpointConfiguration, onCriticalError, transportCustomization, settings);
         hostBuilder.UseNServiceBus(endpointConfiguration);
     }
 
-    internal static void ConfigureEndpoint(EndpointConfiguration config, Func<ICriticalErrorContext, CancellationToken, Task> onCriticalError, ITransportCustomization transportCustomization, Settings settings)
+    static void ConfigureEndpoint(EndpointConfiguration config, Func<ICriticalErrorContext, CancellationToken, Task> onCriticalError, ITransportCustomization transportCustomization, Settings settings)
     {
         if (!string.IsNullOrWhiteSpace(settings.LicenseFileText))
         {
@@ -107,8 +108,6 @@ public static class HostApplicationBuilderExtensions
         config.AddDeserializer<TaggedLongValueWriterOccurrenceSerializerDefinition>();
         config.Pipeline.Register(typeof(MessagePoolReleasingBehavior), "Releases pooled message.");
         config.EnableFeature<QueueLength.QueueLength>();
-
-        config.EnableFeature<LicenseCheckFeature>();
     }
 
     static Func<QueueLengthStore, IProvideQueueLength> QueueLengthProviderBuilder(string connectionString,
