+++ src/ServiceControl.Monitoring/Licensing/ActiveLicense.cs
@@ -1,0 +1,27 @@
+﻿namespace ServiceControl.Monitoring.Licensing
+{
+    using global::ServiceControl.LicenseManagement;
+    using NServiceBus.Logging;
+
+    public class ActiveLicense
+    {
+        public ActiveLicense() => Refresh();
+
+        public bool IsValid { get; set; }
+
+        internal LicenseDetails Details { get; set; }
+
+        public void Refresh()
+        {
+            Logger.Debug("Refreshing ActiveLicense");
+
+            var detectedLicense = LicenseManager.FindLicense();
+
+            IsValid = !detectedLicense.Details.HasLicenseExpired();
+
+            Details = detectedLicense.Details;
+        }
+
+        static readonly ILog Logger = LogManager.GetLogger(typeof(ActiveLicense));
+    }
+}
