+++ src/ServiceControl.Monitoring/Licensing/LicenseCheckServiceCollectionExtensions.cs
@@ -1,0 +1,13 @@
+﻿namespace ServiceControl.Monitoring.Licensing
+{
+    using Microsoft.Extensions.DependencyInjection;
+
+    static class LicenseCheckServiceCollectionExtensions
+    {
+        public static void AddLicenseCheck(this IServiceCollection services)
+        {
+            services.AddSingleton<ActiveLicense>();
+            services.AddHostedService<LicenseCheckHostedService>();
+        }
+    }
+}
