--- src/ServiceControl/Licensing/LicenseCheckHostBuilderExtensions.cs
@@ -1,16 +1,0 @@
-﻿namespace Particular.ServiceControl.Licensing
-{
-    using Microsoft.Extensions.DependencyInjection;
-    using Microsoft.Extensions.Hosting;
-
-    static class LicenseCheckHostBuilderExtensions
-    {
-        public static IHostApplicationBuilder AddLicenseCheck(this IHostApplicationBuilder hostBuilder)
-        {
-            var services = hostBuilder.Services;
-            services.AddSingleton<ActiveLicense>();
-            services.AddHostedService<LicenseCheckHostedService>();
-            return hostBuilder;
-        }
-    }
-}
